Move waterfall to utils/charts.py
