+++ policy_engine_uk/populations/charts.py
@@ -1,10 +1,9 @@
 from policy_engine_uk.populations.metrics import poverty_rate, pct_change
 from policy_engine_uk.utils.charts import (
     format_fig,
+    waterfall,
     BLUE,
     GRAY,
-    DARK_BLUE,
-    WHITE,
     DARK_GRAY,
     LIGHT_GRAY,
     LIGHT_GREEN,
@@ -115,53 +114,6 @@ def add_zero_line(fig):
         line=dict(color="grey", width=1, dash="dash"),
     )
     return fig
-
-
-def waterfall(values, labels, gain_label="Revenue", loss_label="Spending"):
-    final_color = DARK_BLUE
-
-    def amount_reform_type(amount, reform, type):
-        return pd.DataFrame({"Amount": amount, "Reform": reform, "Type": type})
-
-    if len(labels) == 0:
-        df = amount_reform_type([], [], [])
-    else:
-        df = amount_reform_type(values, labels, "")
-        if len(df) != 0:
-            order = np.where(
-                df.Amount >= 0, -np.log(df.Amount), 1e2 - np.log(-df.Amount)
-            )
-            df = df.set_index(order).sort_index().reset_index(drop=True)
-            df["Type"] = np.where(df.Amount >= 0, gain_label, loss_label)
-            base = np.array([0] + list(df.Amount.cumsum()[:-1]))
-            final_value = df.Amount.cumsum().values[-1]
-            if final_value >= 0:
-                final_color = DARK_BLUE
-            else:
-                final_color = DARK_GRAY
-            df = pd.concat(
-                [
-                    amount_reform_type(base, df.Reform, ""),
-                    df,
-                    amount_reform_type([final_value], ["Final"], ["Final"]),
-                ]
-            )
-        else:
-            df = amount_reform_type([], [], [])
-    fig = px.bar(
-        df.round(),
-        x="Reform",
-        y="Amount",
-        color="Type",
-        barmode="stack",
-        color_discrete_map={
-            gain_label: BLUE,
-            loss_label: GRAY,
-            "": WHITE,
-            "Final": final_color,
-        },
-    )
-    return format_fig(fig, show=False)
 
 
 def total_income(sim):
