+++ policy_engine_uk/situations/charts.py
@@ -1,6 +1,5 @@
 from openfisca_uk import IndividualSim
-from policy_engine_uk.utils.charts import format_fig, GRAY, BLUE
-from policy_engine_uk.populations.charts import waterfall
+from policy_engine_uk.utils.charts import format_fig, GRAY, BLUE, waterfall
 import json
 import plotly.express as px
 import pandas as pd
